feat(ava): fix advise problem
