+++ packages/ava/src/data/index.ts
@@ -1,5 +1,4 @@
 export * from './model';
 export * from './shard';
-export * from './types';
 export * from './constants';
 export * from './infer';
