+++ site/examples/advice/advise-summary/demo/common.tsx
@@ -1,13 +1,13 @@
 import React, { useState, useCallback } from 'react';
 
 import ReactDOM from 'react-dom';
-import { Input } from 'antd';
+import { Input, Button } from 'antd';
 import { Advisor } from '@antv/ava';
 import { renderChart } from '@antv/ava-renderer';
 
 const advisor = new Advisor({
   llm: {
-    appId: '202508APgb7V00506760',
+    appId: '202510APxPmo00551539',
     authorization: 'TBox-c4ae8a71224e42baaafb1c01d15395a7',
   },
 });
@@ -40,7 +40,7 @@ const App = () => {
   return (
     <div>
       <Input.TextArea value={JSON.stringify(data)} />
-      <button onClick={advise}>advise</button>
+      <Button onClick={advise}>advise</Button>
       <div>{chart}</div>
     </div>
   );
